requirements.txt
albumentations==1.1.0
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
pytorch-lightning==1.6.3
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
wandb==0.12.16